New module for collecting dependency-based context counts. Combined with existing modules, this gives us a way to build a count matrix. Next step is to perform matrix factorization on this, which I've not implemented yet.

+++ src/python/pimlico/modules/features/vocab_builder/info.py
@@ -14,7 +14,7 @@ class ModuleInfo(BaseModuleInfo):
         },
         "term_max_prop": {
             "help": "Include terms that occur in max this proportion of documents",
-            "type": int,
+            "type": float,
         },
         "term_limit": {
             "help": "Limit vocab size to this number of most common entries (after other filters)",
@@ -26,7 +26,7 @@ class ModuleInfo(BaseModuleInfo):
         },
         "feature_max_prop": {
             "help": "Include features that occur in max this proportion of documents",
-            "type": int,
+            "type": float,
         },
         "feature_limit": {
             "help": "Limit vocab size to this number of most common entries (after other filters)",
